refactor(api): nasty code in the api
